%  File ergm/man/networksis-internal.Rd
%  Part of the statnet package, http://statnetproject.org
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnetproject.org/attribution
%
% Copyright 2003 Mark S. Handcock, University of Washington
%                David R. Hunter, Penn State University
%                Carter T. Butts, University of California - Irvine
%                Steven M. Goodreau, University of Washington
%                Martina Morris, University of Washington
% Copyright 2007 The statnet Development Team
\name{networksis-internal}
\title{Internal networksis Objects}
% terms
\alias{networksis.slave}
\alias{sis.conjugate}
\alias{sis.knots}
\alias{sis.rowsort}
\alias{sis.simulate}
%
% These need to be documented!
%
\alias{InitErgm.coincidences}
\alias{simulate.network}
\alias{simulate.numeric}

\description{
  Internal networksis functions.
}
%\usage{
%anova.networksis(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{networksis, networksis-package, ergm, statnet}
\keyword{internal}
