%  File networksis/man/networksis-internal.Rd
%  Part of the statnet package, http://statnet.org
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnet.org/attribution
%
%  Copyright 2012 the statnet development team
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{networksis-internal}
\title{Internal networksis Objects}
% terms
\alias{simulate.network}
\description{
  Internal networksis functions.
}
\details{
  Most of these functions are not to be called by the user.
}
\keyword{internal}
